\name{abalone}
\alias{abalone}
\docType{data}
\title{
data name/kind ... ~~
}
\description{
A concise (1-5 lines) description of the dataset. ~~
}
\usage{data(abalone)}
\format{
  The format is:
 chr "abalone"
}
\details{
If necessary, more details than the __description__ above ~~
}
\source{
reference to a publication or URL from which the data were obtained ~~
}
\references{
possibly secondary sources and usages ~~
}
\examples{
data(abalone)
## maybe str(abalone) ; plot(abalone) ...
}
\keyword{datasets}
