\name{clean.madlib.temp}
\alias{clean.madlib.temp}

\title{
  Delete all the result tables created during calculations of
  MADlib 
}
\description{
 
}

\usage{
clean.madlib.temp(conn.id = 1)
}

\arguments{
  \item{conn.id}{

  }
}

\details{

}

\author{
  Author: Predictive Analytics Team at Pivotal Inc. \email{user@madlib.net}
  
  Maintainer: Hai Qian \email{hqian@gopivotal.com}, Predictive Analytics Team at Pivotal
  Inc. \email{user@madlib.net}
}

\seealso{
  \code{\link{madlib.arima}} creates three tables with names starting
  with "__madlib_temp_" when it fits ARIMA model to time series
}

\examples{
\dontrun{


}
}

\keyword{utilities}
\keyword{ database }
\keyword{ misc }