\name{conn.eql}
\alias{conn.eql}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Check whether two connections are equal
}
\description{
  A concise (1-5 lines) description of what the function does. ~~
}
\usage{
conn.eql(conn.id1, conn.id2)
}

\arguments{
  \item{conn.id1}{
     Describe \code{conn.id1} here~~
}
  \item{conn.id2}{
     Describe \code{conn.id2} here~~
}
}
\details{
 If necessary, more details than the description above ~~
}
\value{
~Describe the value returned
  If it is a LIST, use
  \item{conn.id1}{Connection 1}
  \item{conn.id2}{Connection 2}
}
\references{
put references to the literature/web site here ~
}
\author{
who you are~~
}
\note{
further notes~~
}

\seealso{
objects to See Also as \code{\link{help}}, ~~~
}
\examples{
\dontrun{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ kwd1 }
\keyword{ kwd2 }% __ONLY ONE__ keyword per line
