\name{madlib.summary}
\alias{madlib.summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function to do ... ~~
}
\description{
 A concise (1-5 lines) description of what the function does. ~~
}
\usage{
madlib.summary(x, target.cols = NULL, grouping.cols = NULL, get.distinct = TRUE, get.quartiles = TRUE, ntile = NULL, n.mfv = 10, estimate = TRUE, interactive = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Describe \code{x} here~~
}
  \item{target.cols}{
Describe \code{target.cols} here~~
}
  \item{grouping.cols}{
Describe \code{grouping.cols} here~~
}
  \item{get.distinct}{
Describe \code{get.distinct} here~~
}
  \item{get.quartiles}{
Describe \code{get.quartiles} here~~
}
  \item{ntile}{
Describe \code{ntile} here~~
}
  \item{n.mfv}{
Describe \code{n.mfv} here~~
}
  \item{estimate}{
Describe \code{estimate} here~~
}
  \item{interactive}{
Describe \code{interactive} here~~
}
}
\details{
If necessary, more details than the description above ~~
}
\value{
~Describe the value returned
If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
put references to the literature/web site here ~
}
\author{
who you are~~
}
\note{
further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
objects to See Also as \code{\link{help}}, ~~~
}
\examples{
\dontrun{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"madlib.summary"
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ kwd1 }
\keyword{ kwd2 }% __ONLY ONE__ keyword per line
