\name{GUI}
\alias{PivotalR}
\alias{pivotalr}

\title{
  Graphical interface for PivotalR based upon shiny
}

\description{
  This function launches a shiny server which provides a graphical
  interface for PivotalR. Press Ctrl+c to stop the shiny server.
}

\usage{
PivotalR()

pivotalr()
}

\details{
  The graphical interface for PivotalR is very easy to use. Just follow
  the instructions on screen. The GUI is still at a very early stage and
  has only very limited functionality. We will add more functionalities
  into the GUI in the future versions.
}

\author{
  Author: Predictive Analytics Team at Pivotal Inc.
  
  Maintainer: Hai Qian, Pivotal Inc. \email{hqian@pivotal.io}
} 

\references{
  [1] RStudio and Inc. (2013). shiny: Web Application Framework for R. R
  package version 0.6.0. \url{http://CRAN.R-project.org/package=shiny}

  [2] shiny website, \url{http://www.rstudio.com/shiny/}
}

\keyword{GUI}
\keyword{utility}
