\name{as.db.data.frame-methods}
\docType{methods}
\alias{as.db.data.frame}
\alias{as.db.data.frame.character}
\alias{as.db.data.frame.data.frame}
\alias{as.db.data.frame.db.Rquery}
\alias{as.db.data.frame,character-method}
\alias{as.db.data.frame,data.frame-method}
\alias{as.db.data.frame,db.Rquery-method}
\title{ Methods for Function \code{as.db.data.frame} in Package \pkg{PivotalR} ~~}
\description{
  Methods for function \code{as.db.data.frame} in package \pkg{PivotalR} ~~
}
\section{Methods}{
\describe{

\item{\code{signature(x = "character")}}{
describe this method here~~
}

\item{\code{signature(x = "data.frame")}}{
Describe this method here~~
}

\item{\code{signature(x = "db.Rquery")}}{
Describe this method here~~
}

}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
