\name{db.Rquery-class}
\Rdversion{1.1}
\docType{class}
\alias{db.Rquery-class}

\title{Class \code{"db.Rquery"}}
\description{
A concise (1-5 lines) description of what the class is. ~~
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("db.Rquery", ...)}.
 describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{.content}:}{Object of class \code{"character"} ~~ }
    \item{\code{.conn.id}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{.col.name}:}{Object of class \code{"character"} ~~ }
    \item{\code{.key}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{db.obj}"}, directly.
}
\section{Methods}{
No methods defined with class "db.Rquery" in the signature.
}
\references{
put references to the literature/web site here~~
}
\author{
who you are~~
}
\note{
further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  objects to See Also
  %as \code{\link{~~fun~~}}, ~~~
%or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("db.Rquery")
}
\keyword{classes}
