\name{db.view-class}
\Rdversion{1.1}
\docType{class}
\alias{db.view-class}

\title{Class \code{"db.view"}}
\description{
 A concise (1-5 lines) description of what the class is. ~~
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("db.view", ...)}.
 describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{.name}:}{Object of class \code{"character"} ~~ }
    \item{\code{.key}:}{Object of class \code{"character"} ~~ }
    \item{\code{.content}:}{Object of class \code{"character"} ~~ }
    \item{\code{.conn.id}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{.col.name}:}{Object of class \code{"character"} ~~ }
    \item{\code{.col.data_type}:}{Object of class \code{"character"} ~~ }
    \item{\code{.col.udt_name}:}{Object of class \code{"character"} ~~ }
    \item{\code{.table.type}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{db.data.frame}"}, directly.
Class \code{"\linkS4class{db.obj}"}, by class "db.data.frame", distance 2.
}
\section{Methods}{
No methods defined with class "db.view" in the signature.
}
\references{
put references to the literature/web site here~~
}
\author{
who you are~~
}
\note{
further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  objects to See Also
  %as \code{\link{~~fun~~}}, ~~~
%or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("db.view")
}
\keyword{classes}
