\name{madlib.glm}
\alias{madlib.glm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function to do ... ~~
}
\description{
 A concise (1-5 lines) description of what the function does. ~~
}
\usage{
madlib.glm(formula, data, family = "gaussian", na.action, control = list(), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{
Describe \code{formula} here~~
}
  \item{data}{
Describe \code{data} here~~
}
\item{family}{
Describe \code{family} here ~~
}
  \item{na.action}{
Describe \code{na.action} here~~
}
  \item{control}{
Describe \code{control} here~~
}
\item{...}{
    Describe ... here
}
}
\details{
If necessary, more details than the description above ~~
}
\value{
Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
put references to the literature/web site here ~
}
\author{
who you are~~
}
\note{
further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## Not run:
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"madlib.glm"
## End (Not run)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ madlib }
\keyword{ wrapper }
