\name{PivotalR-package}
\alias{PivotalR-package}
\docType{package}
\title{
  Wrapper for MADlib, an open-source library for scalable in-database
  analytics
}
\description{
  The package provides a wrapper for MADlib. MADlib is an open-source
  library for scalable in-database analytics. It provides data-parallel
  implementations of mathematical, statistical and machine-learning
  methods for structured and unstructured data.
}
\details{
\tabular{ll}{
Package: \tab PivotalR\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-04-01\cr
License: \tab GPL (>= 2)\cr
Depends: \tab methods, DBI, RPostgreSQL\cr
}
An overview of how to use the package, including the most important 
 functions 
}
\author{
  Author: Predictive Analytics Team at Pivotal
  Inc. \email{user@madlib.net},  with contributions
  from Woo Jung \email{Woo.Jung@gopivotal.com}
  
  Maintainer: Predictive Analytics Team at Pivotal
  Inc. \email{user@madlib.net}
}
\references{
  [1] MADlib website, \url{http://madlib.net}
  
  [2] MADlib user docs, \url{http://doc.madlib.net/master}

  [3] MADlib Wiki page, \url{http://github.com/madlib/madlib/wiki}

  [4] MADlib contribution guide,
  \url{https://github.com/madlib/madlib/wiki/Contribution-Guide}

  [5] MADlib on GitHub, \url{https://github.com/madlib/madlib}
}

\keyword{ package }

\seealso{
  \code{\link{madlib.lm}} Linear regression

  \code{\link{madlib.glm}} Linear, logistic and multinomial logistic regressions
}

